hash passwords + two modes + album password check [Not working properly]

+++ jv-data-access/src/main/java/forester/jv/data/entity/User.java
@@ -9,7 +9,7 @@ import java.math.BigInteger;
 @Entity
 @Table(name = "users")
 public class User {
-    private BigInteger userId;
+    private Long userId;
     private String login;
     private String password;
     private String name;
@@ -19,11 +19,11 @@ public class User {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "user_id")
-    public BigInteger getUserId() {
+    public Long getUserId() {
         return userId;
     }
 
-    public void setUserId(BigInteger userId) {
+    public void setUserId(Long userId) {
         this.userId = userId;
     }
