+++ jv-data-access/src/main/java/forester/jv/data/entity/UsersGroupsAccess.java
@@ -11,33 +11,33 @@ import java.math.BigInteger;
 @IdClass(UsersGroupsAccessPK.class)
 public class UsersGroupsAccess {
 
-    public UsersGroupsAccess(BigInteger userId, BigInteger groupId) {
+    public UsersGroupsAccess(Long userId, Long groupId) {
         this.userId = userId;
         this.groupId = groupId;
     }
 
     public UsersGroupsAccess(){}
 
-    private BigInteger userId;
-    private BigInteger groupId;
+    private Long userId;
+    private Long groupId;
 
     @Id
     @Column(name = "user_id")
-    public BigInteger getUserId() {
+    public Long getUserId() {
         return userId;
     }
 
-    public void setUserId(BigInteger userId) {
+    public void setUserId(Long userId) {
         this.userId = userId;
     }
 
     @Id
     @Column(name = "group_id")
-    public BigInteger getGroupId() {
+    public Long getGroupId() {
         return groupId;
     }
 
-    public void setGroupId(BigInteger groupId) {
+    public void setGroupId(Long groupId) {
         this.groupId = groupId;
     }
