+++ jv-data-access/src/main/java/forester/jv/data/entity/UsersGroupsAccessPK.java
@@ -9,26 +9,26 @@ import java.math.BigInteger;
  * Created by FORESTER on 24.09.17.
  */
 public class UsersGroupsAccessPK implements Serializable {
-    private BigInteger userId;
-    private BigInteger groupId;
+    private Long userId;
+    private Long groupId;
 
     @Column(name = "user_id")
     @Id
-    public BigInteger getUserId() {
+    public Long getUserId() {
         return userId;
     }
 
-    public void setUserId(BigInteger userId) {
+    public void setUserId(Long userId) {
         this.userId = userId;
     }
 
     @Column(name = "group_id")
     @Id
-    public BigInteger getGroupId() {
+    public Long getGroupId() {
         return groupId;
     }
 
-    public void setGroupId(BigInteger groupId) {
+    public void setGroupId(Long groupId) {
         this.groupId = groupId;
     }
